Fix deploy
