+++ src/app/[locale]/sign-in/page.tsx
@@ -1,30 +1,10 @@
-'use client';
-
-import { ENUM_AUTH_FORM_VARIANT } from '../../components/auth/enums';
-import { AuthPage } from '../../components/auth/AuthPage';
-import { useRouter, useSearchParams } from 'next/navigation';
-import { useUserSession } from '../../contexts/auth/hooks/useUserSession';
-import { useEffect } from 'react';
+import { Suspense } from 'react';
+import SuspendedPage from './SuspendedPage';
 
 export default function SignInPage() {
-  const router = useRouter();
-  const prevRoute = useSearchParams().get('from');
-  const { user } = useUserSession();
-
-  useEffect(() => {
-    if (user) {
-      router.push(prevRoute ?? '/');
-    }
-  }, [user]);
-
-  const handleSuccess = () => {
-    router.push(prevRoute ?? '/');
-  };
-
   return (
-    <AuthPage
-      variant={ENUM_AUTH_FORM_VARIANT.SIGNIN}
-      onSuccess={handleSuccess}
-    />
+    <Suspense fallback={<div>Loading...</div>}>
+      <SuspendedPage />
+    </Suspense>
   );
 }
