+++ src/app/page.tsx
@@ -1,5 +1,5 @@
 import { redirect } from 'next/navigation';
-import { defaultLocale } from '../intl/config';
+import { defaultLocale } from '../i18n/routing';
 
 export default function RootPage() {
   redirect(defaultLocale);
