+++ src/i18n/routing.ts
@@ -1,12 +1,14 @@
 import { defineRouting } from 'next-intl/routing';
 import { createNavigation } from 'next-intl/navigation';
 
+export const locales = ['fr', 'en'];
+export const defaultLocale = 'fr';
 export const routing = defineRouting({
   // A list of all locales that are supported
-  locales: ['fr', 'en'],
+  locales,
 
   // Used when no locale matches
-  defaultLocale: 'fr'
+  defaultLocale
 });
 
 // Lightweight wrappers around Next.js' navigation APIs
