+++ src/middleware.ts
@@ -1,15 +1,10 @@
 import createMiddleware from 'next-intl/middleware';
-import { pathnames, locales, localePrefix, defaultLocale } from './intl/config';
 import { NextRequest } from 'next/server';
+import { routing } from './i18n/routing';
 
 export default async function middleware(request: NextRequest) {
-  const handleI18nRouting = createMiddleware({
-    defaultLocale,
-    locales,
-    pathnames,
-    localePrefix
-  });
-  const response = handleI18nRouting(request);
+  const intlMiddleware = createMiddleware(routing);
+  const response = intlMiddleware(request);
   return response;
 }
