--- src/navigation.ts
@@ -1,9 +1,0 @@
-import { createLocalizedPathnamesNavigation } from 'next-intl/navigation';
-import { locales, pathnames, localePrefix } from './intl/config';
-
-export const { Link, redirect, usePathname, useRouter } =
-  createLocalizedPathnamesNavigation({
-    locales,
-    pathnames,
-    localePrefix
-  });
